beta3 mass commit
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@27 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ tests/cv/src/akmeans.cpp
@@ -69,7 +69,7 @@ float distance(float* vector_1, float *vector_2, int VecSize)
 
 //returns number of made iterations
 int _real_kmeans( int numClusters, float **sample, int numSamples, 
-                   int VecSize, int* a_class, float eps, int iter )
+                   int VecSize, int* a_class, double eps, int iter )
 
 {                            
   int     i, k, n;
